+++ minix-plugin/api-paper/src/main/kotlin/dev/racci/minix/api/data/enums/LiquidType.kt
@@ -3,6 +3,7 @@ package dev.racci.minix.api.data.enums
 import org.bukkit.Material
 import org.bukkit.block.Block
 import org.bukkit.block.data.Waterlogged
+import java.util.concurrent.ConcurrentHashMap
 
 public enum class LiquidType {
     WATER,
@@ -10,17 +11,25 @@ public enum class LiquidType {
     NON;
 
     public companion object {
+        private val typeCache = ConcurrentHashMap<Material, LiquidType>()
 
-        public fun convert(block: Block): LiquidType = when (block.type) {
-            Material.WATER -> WATER
-            Material.LAVA -> LAVA
-            else -> if (block is Waterlogged && block.isWaterlogged) WATER else NON
+        public fun convert(block: Block): LiquidType {
+            val waterlogged = block.blockData as? Waterlogged ?: return convert(block.type)
+            return if (waterlogged.isWaterlogged) WATER else NON
         }
 
-        public fun convert(bucket: Material): LiquidType = when (bucket) {
-            Material.WATER_BUCKET -> WATER
-            Material.LAVA_BUCKET -> LAVA
-            else -> NON
+        public fun convert(type: Material): LiquidType = typeCache.computeIfAbsent(type) {
+            when (type) {
+                Material.LAVA,
+                Material.LAVA_BUCKET -> LAVA
+                Material.WATER,
+                Material.WATER_BUCKET,
+                Material.BUBBLE_COLUMN,
+                Material.KELP_PLANT,
+                Material.SEAGRASS,
+                Material.TALL_SEAGRASS -> WATER
+                else -> NON
+            }
         }
 
         public val Block.liquidType: LiquidType get() = convert(this)