+++ minix-plugin/api-paper/src/main/kotlin/dev/racci/minix/api/flow/Extensions.kt
@@ -10,6 +10,7 @@ import org.bukkit.entity.Player
 import org.bukkit.event.EventPriority
 import org.bukkit.event.Listener
 import org.bukkit.event.player.PlayerEvent
+import kotlin.reflect.KClass
 
 /**
  * Creates an event flow for [PlayerEvent] that auto filter for only events from [player].
@@ -20,7 +21,16 @@ public inline fun <reified T : PlayerEvent> WithPlugin<*>.playerEventFlow(
     ignoreCancelled: Boolean = false,
     channel: Channel<T> = Channel(Channel.CONFLATED),
     listener: Listener = SimpleKListener(plugin)
-): Flow<T> = plugin.playerEventFlow(player, priority, ignoreCancelled, channel, listener)
+): Flow<T> = plugin.playerEventFlow(T::class, player, priority, ignoreCancelled, channel, listener)
+
+public inline fun <T : PlayerEvent> WithPlugin<*>.playerEventFlow(
+    kClass: KClass<T>,
+    player: Player,
+    priority: EventPriority = EventPriority.NORMAL,
+    ignoreCancelled: Boolean = false,
+    channel: Channel<T> = Channel(Channel.CONFLATED),
+    listener: Listener = SimpleKListener(plugin)
+): Flow<T> = plugin.playerEventFlow(kClass, player, priority, ignoreCancelled, channel, listener)
 
 public inline fun <reified T : PlayerEvent> MinixPlugin.playerEventFlow(
     player: Player,
@@ -28,7 +38,17 @@ public inline fun <reified T : PlayerEvent> MinixPlugin.playerEventFlow(
     ignoreCancelled: Boolean = false,
     channel: Channel<T> = Channel(Channel.CONFLATED),
     listener: Listener = SimpleKListener(this)
-): Flow<T> = playerEventFlow(player, this, priority, ignoreCancelled, channel, listener)
+): Flow<T> = playerEventFlow(T::class, player, this, priority, ignoreCancelled, channel, listener)
+
+public inline fun <T : PlayerEvent> MinixPlugin.playerEventFlow(
+    kClass: KClass<T>,
+    player: Player,
+    priority: EventPriority = EventPriority.NORMAL,
+    ignoreCancelled: Boolean = false,
+    channel: Channel<T> = Channel(Channel.CONFLATED),
+    listener: Listener = SimpleKListener(this)
+): Flow<T> = playerEventFlow(kClass, player, this, priority, ignoreCancelled, channel, listener)
+
 
 public inline fun <reified T : PlayerEvent> playerEventFlow(
     player: Player,
@@ -37,5 +57,15 @@ public inline fun <reified T : PlayerEvent> playerEventFlow(
     ignoreCancelled: Boolean = false,
     channel: Channel<T> = Channel(Channel.CONFLATED),
     listener: Listener = SimpleKListener(plugin)
-): Flow<T> = eventFlow(T::class, plugin, player, priority, ignoreCancelled, false, channel, listener)
+): Flow<T> = playerEventFlow(T::class, player, plugin, priority, ignoreCancelled, channel, listener)
+
+public inline fun <T : PlayerEvent> playerEventFlow(
+    kClass: KClass<T>,
+    player: Player,
+    plugin: MinixPlugin,
+    priority: EventPriority = EventPriority.NORMAL,
+    ignoreCancelled: Boolean = false,
+    channel: Channel<T> = Channel(Channel.CONFLATED),
+    listener: Listener = SimpleKListener(plugin)
+): Flow<T> = eventFlow(kClass, plugin, player, priority, ignoreCancelled, false, channel, listener)
     .filter { it.player.name == player.name }