+++ minix-plugin/core-common/src/main/kotlin/dev/racci/minix/core/coroutine/MinixCoroutineExceptionHandler.kt
@@ -4,6 +4,7 @@ import dev.racci.minix.api.events.plugin.CaughtCoroutineExceptionEvent
 import dev.racci.minix.api.plugin.MinixPlugin
 import dev.racci.minix.api.utils.getKoin
 import dev.racci.minix.flowbus.FlowBus
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineExceptionHandler
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.Scoped
@@ -22,6 +23,8 @@ internal fun MinixCoroutineExceptionHandler(
             plugin.logger.debug { "Received exception while plugin was disabled: $exception" }
             return
         }
+
+        if (exception is CancellationException) return
 
         plugin.launch {
             val event = CaughtCoroutineExceptionEvent(plugin, exception)