+++ minix-plugin/core-paper/src/main/kotlin/dev/racci/minix/core/plugin/DummyLoader.kt
@@ -10,6 +10,7 @@ import org.bukkit.plugin.Plugin
 import org.bukkit.plugin.PluginDescriptionFile
 import org.bukkit.plugin.SimplePluginManager
 import org.bukkit.plugin.java.PluginClassLoader
+import java.lang.ref.WeakReference
 import kotlin.reflect.KMutableProperty1
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.jvm.isAccessible
@@ -27,7 +28,8 @@ public class DummyLoader {
         // These values need to be null when a JavaPlugin class creates a new instance on the classloader.
         classLoader.setValue<PluginClassLoader, Plugin?>("plugin", null)
         classLoader.setValue<PluginClassLoader, Plugin?>("pluginInit", null)
-        val minix = MinixImpl()
+
+        val minix = MinixImpl(WeakReference(initPlugin))
         classLoader.setValue<PluginClassLoader, Plugin?>("plugin", minix)
         classLoader.setValue<PluginClassLoader, Plugin?>("pluginInit", minix)
         minix.onLoad()
@@ -36,29 +38,28 @@ public class DummyLoader {
             // Delay for a nanosecond to allow the method to return, otherwise the plugin won't have been added to these lists/maps yet.
             Thread.sleep(0, 1)
 
-            val plugins = getValue<SimplePluginManager, ArrayList<Plugin>>("plugins", pluginManager as SimplePluginManager)
+            val plugins = (pluginManager as SimplePluginManager).getValue<SimplePluginManager, ArrayList<Plugin>>("plugins")
             plugins.remove(initPlugin)
             plugins.add(minix)
-            val lookupNames = getValue<SimplePluginManager, HashMap<String, Plugin>>("lookupNames", pluginManager as SimplePluginManager)
+            val lookupNames = (pluginManager as SimplePluginManager).getValue<SimplePluginManager, HashMap<String, Plugin>>("lookupNames")
             lookupNames.remove(initPlugin.name.lowercase())
             lookupNames[minix.name.lowercase()] = minix
         }.start()
     }
 
-    private inline fun <reified T : Any, R : Any> getValue(
-        field: String,
-        obj: T
-    ): R = T::class.memberProperties.findKCallable(field, false)
-        .tap { it.isAccessible = true }
-        .map { it.get(obj).castOrThrow<R>() to it }
-        .tap { it.second.isAccessible = false }
-        .orNull()!!.first
+    internal companion object {
+        inline fun <reified T : Any, R : Any> T.getValue(field: String): R = T::class.memberProperties.findKCallable(field, false)
+            .tap { it.isAccessible = true }
+            .map { it.get(this).castOrThrow<R>() to it }
+            .tap { it.second.isAccessible = false }
+            .orNull()!!.first
 
-    private inline fun <reified T : Any, reified V> T.setValue(
-        field: String,
-        value: V
-    ) = T::class.memberProperties.findKCallable(field)
-        .tap { it.isAccessible = true }
-        .filterIsInstance<KMutableProperty1<*, *>>()
-        .map { it.setter.call(this, value) }
+        inline fun <reified T : Any, reified V> T.setValue(
+            field: String,
+            value: V
+        ) = T::class.memberProperties.findKCallable(field)
+            .tap { it.isAccessible = true }
+            .filterIsInstance<KMutableProperty1<*, *>>()
+            .map { it.setter.call(this, value) }
+    }
 }